SaintPetersburg_QuestBot/requirements.txt
telebot
random
pyyaml
emoji
ultralytics
Pillow
yolo
Pyrebase4
pytorch==2.3.1


